+++ neoforge/src/main/java/com/diontryban/ash_api/resources/ResourceLoaderNeoForge.java
@@ -20,22 +20,22 @@
 package com.diontryban.ash_api.resources;
 
 import com.diontryban.ash_api.AshApi;
-import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.packs.PackType;
 import net.minecraft.server.packs.resources.PreparableReloadListener;
-import net.minecraft.server.packs.resources.ReloadableResourceManager;
+import net.neoforged.neoforge.client.event.AddClientReloadListenersEvent;
 import net.neoforged.neoforge.common.NeoForge;
 import net.neoforged.neoforge.event.AddServerReloadListenersEvent;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.Queue;
+import java.util.concurrent.ConcurrentLinkedQueue;
 
 @ApiStatus.Internal
 public final class ResourceLoaderNeoForge extends ResourceLoader {
-    private static final List<PreparableReloadListener> SERVER_RELOAD_LISTENERS = new ArrayList<>();
+    private static final Queue<PreparableReloadListener> CLIENT_RELOAD_LISTENERS = new ConcurrentLinkedQueue<>();
+    private static final Queue<PreparableReloadListener> SERVER_RELOAD_LISTENERS = new ConcurrentLinkedQueue<>();
 
     private final PackType type;
 
@@ -55,14 +55,13 @@ public final class ResourceLoaderNeoForge extends ResourceLoader {
     @Override
     public void registerReloadListener(@NotNull PreparableReloadListener reloadListener) {
         if (type == PackType.CLIENT_RESOURCES) {
-            // TODO: Test if this breaks on a dedicated server.
-            var resourceManager = Minecraft.getInstance().getResourceManager();
-            if (resourceManager instanceof ReloadableResourceManager) {
-                ((ReloadableResourceManager) resourceManager).registerReloadListener(reloadListener);
+            if (CLIENT_RELOAD_LISTENERS.isEmpty()) {
+                NeoForge.EVENT_BUS.addListener(this::onAddClientReloadListener);
             }
+            CLIENT_RELOAD_LISTENERS.add(reloadListener);
         } else if (type == PackType.SERVER_DATA) {
             if (SERVER_RELOAD_LISTENERS.isEmpty()) {
-                NeoForge.EVENT_BUS.addListener(this::onAddReloadListener);
+                NeoForge.EVENT_BUS.addListener(this::onAddServerReloadListener);
             }
             SERVER_RELOAD_LISTENERS.add(reloadListener);
         } else {
@@ -70,7 +69,13 @@ public final class ResourceLoaderNeoForge extends ResourceLoader {
         }
     }
 
-    private void onAddReloadListener(AddServerReloadListenersEvent event) {
+    private void onAddClientReloadListener(AddClientReloadListenersEvent event) {
+        for (PreparableReloadListener reloadListener : CLIENT_RELOAD_LISTENERS) {
+            event.addListener(ResourceLocation.parse(reloadListener.getName()), reloadListener);
+        }
+    }
+
+    private void onAddServerReloadListener(AddServerReloadListenersEvent event) {
         for (PreparableReloadListener reloadListener : SERVER_RELOAD_LISTENERS) {
             event.addListener(ResourceLocation.parse(reloadListener.getName()), reloadListener);
         }